- Ajout de la classe Square : permet une utilisation rapide des carreaux du plateau de jeu (exemple avec le chateau dans World() etant donne que nous connaissons leur emplacement)
- Modification de toutes les listes :
j'ai tout remis en List<Monster>, etc et modifiant ton code afin que ca
fonction, syntaxement parlant il me semble que c'est plus correcte (sauf
erreur de ma part)

- Modification dans toute la classe World :
surtout annotation et ordres des fonctions (j'ai essaye de bien ranger
tout ca en faisant quelques modif afin d ameliorer le tout et que ce
soit plus beau)

- Modification de l initialisation du plateau :
J'ai tout mis directement dans les parametre, ca me parraissait mieux,
tu me diras ce que tu en pense.

- et deux trois autres trucs par ci par la

bonne année c:

+++ src/warcraftPath/Patern1.java
@@ -15,7 +15,7 @@ public class Patern1 {
 	public static List<Position> pathconstruct() {
 		List<Position> path = new LinkedList<Position>();
 		for(int i=0; i<24; i++)
-			path.add(new Position(i * World.getSquareWidth() + World.getSquareWidth() / 2, 7 * World.getSquareHeight() + World.getSquareHeight() / 2));
+			path.add(new Position(i * World.getSquareWidth() + World.getSquareWidth() / (float)2.0, 7 * World.getSquareHeight() + World.getSquareHeight() / (float)2.0));
 		return path;
 	}
